move modal view to backend page layout
